Modernize Rails integration
Rails 7.1 adds an :after_routes_loaded hook. Update the railtie to
reload allowed routes when Rails routes change (in development mode).
Since the hook is only available in Rails 7.1, fallback to the existing
:after_initialize hook when the :after_routes_loaded hook is not
defined.

Additionally, match Rails default 404 behavior by returning the contents
of public/404.html when it is present, and routes are not allowed.
